Fixed Sized Bound Queue Caller Runs
Creating a Fixed-sized Thread Pool with a Bounded Queue and the Caller-runs
Saturation Policy.

The caller-runs policy implements a form of throttling that neither discards
tasks nor throws an exception, but instead tries to slow down the flow of new
tasks by pushing some of the work back to the caller. It executes the newly
submitted task not in a pool thread, but in the thread that calls execute.